qtgui: only one definition of (internal) TimePrecisionClass
(cherry picked from commit e1b633a95047fce266e575cbc66a314d31c24491)
