+++ gr-qtgui/lib/TimeDomainDisplayPlot.cc
@@ -11,6 +11,7 @@
 #ifndef TIME_DOMAIN_DISPLAY_PLOT_C
 #define TIME_DOMAIN_DISPLAY_PLOT_C
 
+#include "TimePrecisionClass.h"
 #include <gnuradio/qtgui/TimeDomainDisplayPlot.h>
 
 #include <qwt_legend.h>
@@ -19,25 +20,6 @@
 #include <QColor>
 #include <cmath>
 #include <iostream>
-
-class TimePrecisionClass
-{
-public:
-    TimePrecisionClass(const int timePrecision) { d_timePrecision = timePrecision; }
-
-    virtual ~TimePrecisionClass() {}
-
-    virtual unsigned int getTimePrecision() const { return d_timePrecision; }
-
-    virtual void setTimePrecision(const unsigned int newPrecision)
-    {
-        d_timePrecision = newPrecision;
-    }
-
-protected:
-    unsigned int d_timePrecision;
-};
-
 
 class TimeDomainDisplayZoomer : public QwtPlotZoomer, public TimePrecisionClass
 {
