+++ gr-qtgui/lib/TimeRasterDisplayPlot.cc
@@ -11,6 +11,7 @@
 #ifndef TIMERASTER_DISPLAY_PLOT_C
 #define TIMERASTER_DISPLAY_PLOT_C
 
+#include "TimePrecisionClass.h"
 #include <gnuradio/qtgui/TimeRasterDisplayPlot.h>
 
 #include <gnuradio/qtgui/qtgui_types.h>
@@ -234,24 +235,6 @@ private:
     double d_delta_value;
     double d_ten_scale;
     std::string d_units;
-};
-
-class TimePrecisionClass
-{
-public:
-    TimePrecisionClass(const int timePrecision) { d_timePrecision = timePrecision; }
-
-    virtual ~TimePrecisionClass() {}
-
-    virtual unsigned int getTimePrecision() const { return d_timePrecision; }
-
-    virtual void setTimePrecision(const unsigned int newPrecision)
-    {
-        d_timePrecision = newPrecision;
-    }
-
-protected:
-    unsigned int d_timePrecision;
 };
 
 /***********************************************************************
